permission update and create controller middleware permission

+++ app/Http/Controllers/backend/AssignRoleController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\backend;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
-use DataTables;
+use Yajra\DataTables\DataTables;
 use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 
@@ -24,7 +24,6 @@ class AssignRoleController extends Controller
             $roles = Role::where('name','!=', 'superadmin')->pluck('name')->all();
         }
         if ($request->ajax()) {
-            
             return DataTables::of($users)
                 ->addIndexColumn()
                 ->addColumn('role', function($row) {
@@ -48,9 +47,12 @@ class AssignRoleController extends Controller
         return view('admin.assign-role.index',['roles'=>$roles]);
     }
 
-    public function assignRole(Request $request)
+    public function store(Request $request)
     {
         $user = User::where('email', $request->email)->first();
+        if(!$user){
+            return back()->with('error', 'User Email not found');
+        }
         $user->roles()->detach();
         $user->assignRole($request->role);
         return back()->with('success', 'Role assigned successfully');
